requirements copy.txt
fastapi
uvicorn
transformers
sentencepiece
pywhispercpp
pydantic
langdetect